Worked through compiler errors with Smitty and we have a successful build of EventTask now

+++ MainBrain/src/Model/EventTask/EventTask.cpp
@@ -26,7 +26,7 @@ EventTask::EventTask()
  */
 DeviceStatus EventTask::getDeviceStatus(DeviceName device)
 {
-    return deviceLookupTable[device];
+    return deviceLookupTable[(int)device];
 }
